identity-engineering-sync-app/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
requests>=2.32.0
tenacity>=8.2.3
python-dateutil>=2.9.0
mssql
